KhanAcademyCostars
This file presents a Khan Academy project titled "Project: Famous people". The general goal was to create relational tables of famous people to practice join statements. The join would be used to combine data that could present readable conclusions. I created tables that uses actors, movies they've acted in, and those movies' box office sales, along with relational ids. Line 58 begins the presentation of an INNER JOIN that uses a sub query to call duplicate data relating to unique data in another table.

The main point of this showcase is to demonstrate a problem I asked, "How would I search for costars?" or in general, "How do I purposely call duplicate values (such as a country name or city) and join it to unique relational data.  I used the movie "Glass" starring Bruce Willis and Anya Taylor-Joy as the answer I wanted to build my query towards.  Actors are stored in the table "moviestars" and movie titles are in the table "movies".

Thought Process:
With respect to execution speed, I tested simple statements first. While the DISTINCT statement can find unique data, it can't search for only duplicated data. Simple GROUP BY  or HAVING COUNT() statements omits or joins data together in a manner that is not usable depending how it's built. I used Google to search for similar scenarios which led to Stack Overflow. I found a sub-query within the ON statement could accurately pull this answer.

Ultimately, an INNER JOIN's ON statement used the relational table ids to usefully join actor and movie title data while an AND statement connected a subquery that used the COUNT of movie titles to call a duplicate movie title more than once joined with two different movie stars. 

Relating this to a real world scenario, this query could call data for a particular hot selling product and lists the unique companies or regions this product is being bought in or shipped.